Move =~ to NumericMatchers, add some scaladocs
